Add E2E test random recipe integrate in RecipeController file
